fix: 清掉 PHPStan 基线里的 30 处告警
上一个提交把仓库里既有的 30 处告警放进了 phpstan-baseline.neon。现在逐条
修掉，基线文件删除，level 5 全量零告警。

真实的类型问题：
- CheckTrafficExceeded：Redis::scard() 在 phpredis 的流式模式下会返回
  Redis 对象或 false，直接 `<= 0` 比较在那两种情况下是类型错误，加 (int)。
- DeviceStateService：hgetall 取回的时间戳是字符串，`$now - $timestamp`
  碰上非数字串会 TypeError，两处都强转 int。
- KnowledgeController：preg_replace 失败返回 null，会把按引用传入的
  string $body 打成 null，补 `?? $body`。
- NodeWebSocketServer：option() 给的是 string|null，NodeWorker 要 string+int。
- ManageController：Server::$show 是 bool，复制节点时赋的却是 0。

多余的写法（行为不变）：
- `$x?->y ?? $z` 里的 ?- > 是多余的，?? 本身就吞掉「在 null 上取属性」，
  三处改回 ->（MailTemplateController ×2、V2 UserController ×1）。
- General.php ×3：`if ($extra = ...)` 已经要求真值，里面再 !empty($extra)。
- QuantumultX buildAnyTLS：数组每项都是非空字符串，array_filter 过滤不掉
  任何东西（其余几处的 array_filter 确实有用，保留）。
- SingBox::loadConfig：subscribe_template() 声明是 ?string，is_array()
  永远为假，那个分支走不到。
- AppController：`$config = $config ?? [];` 上一行刚赋过值。
- MailService：正则第 1 组是必配的，`$m[1] ?? ''` 的兜底走不到。
- Helper::randomPort：explode 至少返回一个元素，`$portRange[0] ?? 0` 同理。

注解不准：
- TicketMessage 的 `@property-read Ticket $ticket` 写成了非空，于是两个
  访问器里的判空看着像废话。工单被删时它确实是 null（实测孤儿消息两个
  访问器都返回 false），改成可空而不是删掉判断。
- UserController::changePassword：Sanctum 把 currentAccessToken() 标成
  非空，实际没有 token 时是 null，这个分支必须留着——null 时该清掉全部
  token。加 @var 说明实情。
- AdminAuditLog::admin() 缺返回类型，larastan 认不出它是关联，
  with('admin:id,email') 被判成不存在的关联。
- ServerService::getAvailableUsers 的 @param 写的是不存在的 $groupIds。

配置层面：
- Workerman 的 ConnectionInterface 标了 #[AllowDynamicProperties]，在连接
  对象上挂 nodeId / machineId 是它自己的用法（实测不触发 PHP 8.2 的动态
  属性弃用）。把这个事实通过 universalObjectCratesClasses 告诉 PHPStan，
  而不是去重构一个跑着的 WebSocket 服务，也不是压制告警。

删除 app/Scope/FilterScope.php：全仓库零引用的死代码。

验证：92 个测试全绿；另外把改动过的纯函数（协议构造、占位符渲染、
randomPort）改动前后各跑一遍对比，输出逐字节一致，只有 CSPRNG 的抽样不同。

+++ app/Console/Commands/CheckTrafficExceeded.php
@@ -15,7 +15,7 @@ class CheckTrafficExceeded extends Command
 
     public function handle()
     {
-        $count = Redis::scard('traffic:pending_check');
+        $count = (int) Redis::scard('traffic:pending_check');
         if ($count <= 0) {
             return;
         }
